Remove legacy APIs.
