--- pkg/handlers/notification_disabled.go
@@ -1,37 +1,0 @@
-/****************************************************************************
- * Copyright 2020, Optimizely, Inc. and contributors                        *
- *                                                                          *
- * Licensed under the Apache License, Version 2.0 (the "License");          *
- * you may not use this file except in compliance with the License.         *
- * You may obtain a copy of the License at                                  *
- *                                                                          *
- *    http://www.apache.org/licenses/LICENSE-2.0                            *
- *                                                                          *
- * Unless required by applicable law or agreed to in writing, software      *
- * distributed under the License is distributed on an "AS IS" BASIS,        *
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. *
- * See the License for the specific language governing permissions and      *
- * limitations under the License.                                           *
- ***************************************************************************/
-
-// Package handlers //
-package handlers
-
-import (
-	"net/http"
-)
-
-// DisabledNotificationHandler for disabled notifications
-type DisabledNotificationHandler struct {
-}
-
-// HandleEventSteam implements the http.Handler interface. Returns forbidden status
-func (nh *DisabledNotificationHandler) HandleEventSteam(w http.ResponseWriter, r *http.Request) {
-	http.Error(w, "Streaming not enabled", http.StatusForbidden)
-}
-
-// NewDisabledNotificationHandler is the DisableNotificationHandler factory
-func NewDisabledNotificationHandler() (nh *DisabledNotificationHandler) {
-	return &DisabledNotificationHandler{}
-
-}
