+++ pkg/middleware/cached_test.go
@@ -18,20 +18,18 @@
 package middleware
 
 import (
-	"context"
 	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 
-	"github.com/go-chi/chi"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/suite"
 
+	"github.com/optimizely/go-sdk/pkg/entities"
+
 	"github.com/optimizely/agent/pkg/optimizely"
 	"github.com/optimizely/agent/pkg/optimizely/optimizelytest"
-	"github.com/optimizely/go-sdk/pkg/config"
-	"github.com/optimizely/go-sdk/pkg/entities"
 )
 
 var defaultClient = optimizely.OptlyClient{}
@@ -112,157 +110,6 @@ func (suite *OptlyMiddlewareTestSuite) TestGetClient() {
 	suite.Equal(http.StatusOK, rec.Code)
 }
 
-func (suite *OptlyMiddlewareTestSuite) TestGetUserContext() {
-	attributes := map[string]interface{}{
-		"bool":         true,
-		"str":          "yes",
-		"int":          int64(100),
-		"float":        1.00,
-		"quotedString": "true",
-		"nil":          nil,
-	}
-	expected := optimizely.NewContext("test", attributes)
-
-	mux := chi.NewMux()
-	handler := AssertOptlyContextHandler(suite, expected)
-	mux.With(suite.mw.UserCtx).Get("/{userID}", handler)
-
-	req := httptest.NewRequest("GET", `/test?bool=true&str=yes&int=100&float=1.0&quotedString="true"&nil=`, nil)
-	rec := httptest.NewRecorder()
-
-	mux.ServeHTTP(rec, req)
-	suite.Equal(http.StatusOK, rec.Code)
-}
-
-func (suite *OptlyMiddlewareTestSuite) TestGetUserContextError() {
-	mux := chi.NewMux()
-	handler := ErrorHandler()
-	mux.With(suite.mw.UserCtx).Get("/{userID}/features", handler)
-
-	req := httptest.NewRequest("GET", "//features?foo=true&bar=yes&baz=100", nil)
-	rec := httptest.NewRecorder()
-
-	mux.ServeHTTP(rec, req)
-	suite.Equal(http.StatusBadRequest, rec.Code)
-}
-
-func (suite *OptlyMiddlewareTestSuite) TestFeatureCtxFeatureFound() {
-	mux := chi.NewMux()
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		actual, ok := r.Context().Value(OptlyFeatureKey).(*config.OptimizelyFeature)
-		expected := &config.OptimizelyFeature{
-			ExperimentsMap: make(map[string]config.OptimizelyExperiment),
-			ID:             "1",
-			Key:            "one",
-			VariablesMap:   make(map[string]config.OptimizelyVariable),
-		}
-		suite.True(ok)
-		suite.Equal(expected, actual)
-	}
-	mux.With(suite.mw.FeatureCtx).Get("/features/{featureKey}", handler)
-	req := httptest.NewRequest("GET", "/features/one", nil)
-	req = req.WithContext(context.WithValue(req.Context(), OptlyClientKey, &optimizely.OptlyClient{
-		OptimizelyClient: suite.tc.OptimizelyClient,
-	}))
-	req.Header.Add(OptlySDKHeader, "WITH_TEST_CLIENT")
-	rec := httptest.NewRecorder()
-	mux.ServeHTTP(rec, req)
-	suite.Equal(http.StatusOK, rec.Code)
-}
-
-func (suite *OptlyMiddlewareTestSuite) TestFeatureCtxFeatureNotFound() {
-	mux := chi.NewMux()
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		suite.Fail("FeatureCtx should have returned 404 response without calling handler")
-	}
-	mux.With(suite.mw.FeatureCtx).Get("/features/{featureKey}", handler)
-	req := httptest.NewRequest("GET", "/features/two", nil)
-	req = req.WithContext(context.WithValue(req.Context(), OptlyClientKey, &optimizely.OptlyClient{
-		OptimizelyClient: suite.tc.OptimizelyClient,
-	}))
-	req.Header.Add(OptlySDKHeader, "WITH_TEST_CLIENT")
-	rec := httptest.NewRecorder()
-	mux.ServeHTTP(rec, req)
-	suite.Equal(http.StatusNotFound, rec.Code)
-}
-
-func (suite *OptlyMiddlewareTestSuite) TestFeatureCtxNoURLParam() {
-	mux := chi.NewMux()
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		suite.Fail("FeatureCtx should have returned 400 response without calling handler")
-	}
-	mux.With(suite.mw.FeatureCtx).Get("/features/{featureKey}/", handler)
-	req := httptest.NewRequest("GET", "/features//", nil)
-	req = req.WithContext(context.WithValue(req.Context(), OptlyClientKey, &optimizely.OptlyClient{
-		OptimizelyClient: suite.tc.OptimizelyClient,
-	}))
-	req.Header.Add(OptlySDKHeader, "WITH_TEST_CLIENT")
-	rec := httptest.NewRecorder()
-	mux.ServeHTTP(rec, req)
-	suite.Equal(http.StatusBadRequest, rec.Code)
-}
-
-func (suite *OptlyMiddlewareTestSuite) TestExperimentCtxExperimentFound() {
-	mux := chi.NewMux()
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		actual, ok := r.Context().Value(OptlyExperimentKey).(*config.OptimizelyExperiment)
-		expected := &config.OptimizelyExperiment{
-			ID:  suite.tc.ProjectConfig.ExperimentKeyToIDMap["expOne"],
-			Key: "expOne",
-			VariationsMap: map[string]config.OptimizelyVariation{
-				"variation_1": {
-					ID:           "9999",
-					Key:          "variation_1",
-					VariablesMap: map[string]config.OptimizelyVariable{},
-				},
-			},
-		}
-		suite.True(ok)
-		suite.Equal(expected, actual)
-	}
-	mux.With(suite.mw.ExperimentCtx).Get("/experiments/{experimentKey}", handler)
-	req := httptest.NewRequest("GET", "/experiments/expOne", nil)
-	req = req.WithContext(context.WithValue(req.Context(), OptlyClientKey, &optimizely.OptlyClient{
-		OptimizelyClient: suite.tc.OptimizelyClient,
-	}))
-	req.Header.Add(OptlySDKHeader, "WITH_TEST_CLIENT")
-	rec := httptest.NewRecorder()
-	mux.ServeHTTP(rec, req)
-	suite.Equal(http.StatusOK, rec.Code)
-}
-
-func (suite *OptlyMiddlewareTestSuite) TestExperimentCtxExperimentNotFound() {
-	mux := chi.NewMux()
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		suite.Fail("ExperimentCtx should have returned 404 response without calling handler")
-	}
-	mux.With(suite.mw.ExperimentCtx).Get("/experiments/{experimentKey}", handler)
-	req := httptest.NewRequest("GET", "/experiments/expTwo", nil)
-	req = req.WithContext(context.WithValue(req.Context(), OptlyClientKey, &optimizely.OptlyClient{
-		OptimizelyClient: suite.tc.OptimizelyClient,
-	}))
-	req.Header.Add(OptlySDKHeader, "WITH_TEST_CLIENT")
-	rec := httptest.NewRecorder()
-	mux.ServeHTTP(rec, req)
-	suite.Equal(http.StatusNotFound, rec.Code)
-}
-
-func (suite *OptlyMiddlewareTestSuite) TestExperimentCtxNoURLParam() {
-	mux := chi.NewMux()
-	handler := func(w http.ResponseWriter, r *http.Request) {
-		suite.Fail("ExperimentCtx should have returned 400 response without calling handler")
-	}
-	mux.With(suite.mw.ExperimentCtx).Get("/experiments/{experimentKey}/", handler)
-	req := httptest.NewRequest("GET", "/experiments//", nil)
-	req = req.WithContext(context.WithValue(req.Context(), OptlyClientKey, &optimizely.OptlyClient{
-		OptimizelyClient: suite.tc.OptimizelyClient,
-	}))
-	req.Header.Add(OptlySDKHeader, "WITH_TEST_CLIENT")
-	rec := httptest.NewRecorder()
-	mux.ServeHTTP(rec, req)
-	suite.Equal(http.StatusBadRequest, rec.Code)
-}
-
 // ErrorHandler will panic if reached.
 func ErrorHandler() http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
@@ -275,15 +122,6 @@ func AssertOptlyClientHandler(suite *OptlyMiddlewareTestSuite, expected *optimiz
 		actual, err := GetOptlyClient(r)
 		suite.NoError(err)
 		suite.Same(expected, actual)
-
-	}
-}
-
-func AssertOptlyContextHandler(suite *OptlyMiddlewareTestSuite, expected *optimizely.OptlyContext) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		actual, err := GetOptlyContext(r)
-		suite.NoError(err)
-		suite.Equal(expected, actual)
 
 	}
 }
